Move ExerciseModeType, exerciseMode and allExerciseModes to app store
