+++ src/app/components/navbar/components/settings-popup/settings-popup.actions.ts
@@ -1,7 +1,6 @@
 import { createActionGroup, emptyProps, props } from '@ngrx/store';
 
-import { ExerciseModeType, TestingAgainstType } from '@practice/store';
-import { ThemeType } from '@app/store/app';
+import { ExerciseModeType, ThemeType } from '@app/store/app';
 
 
 export const SettingsPopupAction = createActionGroup(
